Adjusts stationBandpassFilter to new event structure and renames it to electricFieldBandPassFilter
